Add files via upload
This give contain an example to solve the displacement response of SDOF system.
An example is given for better understanding. After going through the example user can give his inputs and observe the change in behaviour of the system with change in various parameters.
Feel free to discuss and suggestions are welcomed.